Change widgets imports after moving.
