+++ spyder/plugins/ipythonconsole/widgets/shell.py
@@ -18,9 +18,11 @@ from spyder.config.gui import config_shortcut
 from spyder.py3compat import to_text_string
 from spyder.utils import programs
 from spyder.utils.ipython.style import create_qss_style, create_style_class
-from spyder.widgets.ipythonconsole import (ControlWidget, DebuggingWidget,
-                                           HelpWidget, NamepaceBrowserWidget,
-                                           PageControlWidget)
+from spyder.plugins.ipythonconsole.widgets import (ControlWidget,
+                                                   DebuggingWidget,
+                                                   HelpWidget,
+                                                   NamepaceBrowserWidget,
+                                                   PageControlWidget)
 
 
 class ShellWidget(NamepaceBrowserWidget, HelpWidget, DebuggingWidget):
